requirements.txt
kafka-python
redis
pyyaml
flask
